# openapi_swagger.json
{"openapi": "3.1.0", "info": {"title": "CitySmart Core", "description": "Headless Urban Simulation Multi-Agent Platform", "version": "1.0.0"}, "paths": {"/api/v1/projects/ingest": {"post": {"tags": ["Urban Ingestion Engine"], "summary": "Ingest Unstructured Urban Document", "description": "Accepts unstructured user prompts and extracts mandatory schema fields using explicit types.GenerateContentConfig.", "operationId": "ingest_unstructured_urban_document_api_v1_projects_ingest_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UnstructuredIngestRequest"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": true, "type": "object", "title": "Response Ingest Unstructured Urban Document Api V1 Projects Ingest Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/projects/amend": {"post": {"tags": ["Urban Ingestion Engine"], "summary": "Amend Project Version", "operationId": "amend_project_version_api_v1_projects_amend_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectAmendmentRequest"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/simulation/{project_id}/state": {"get": {"tags": ["Agent Simulation Core"], "summary": "Get Simulation State", "operationId": "get_simulation_state_api_v1_simulation__project_id__state_get", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SimulationStateResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/simulation/chat": {"post": {"tags": ["Agent Simulation Core"], "summary": "Chat With Citizen Persona", "description": "Allows a user to pitch direct layout changes to a specific persona and get immediate qualitative pushback.", "operationId": "chat_with_citizen_persona_api_v1_simulation_chat_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChatInterrogationRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChatInterrogationResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/simulation/render-concept": {"post": {"tags": ["Agent Simulation Core"], "summary": "Generate Blueprint Render", "operationId": "generate_blueprint_render_api_v1_simulation_render_concept_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConceptRenderRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConceptRenderResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/dashboard/{project_id}": {"get": {"tags": ["Management Dashboard Views"], "summary": "Render Management Analytics Dashboard", "operationId": "render_management_analytics_dashboard_api_v1_dashboard__project_id__get", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"text/html": {"schema": {"type": "string"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/health": {"get": {"tags": ["System Diagnostics"], "summary": "System Health Ping", "operationId": "system_health_ping_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"BlueprintRevision": {"properties": {"original_element": {"type": "string", "title": "Original Element"}, "failure_mode_detected": {"type": "string", "title": "Failure Mode Detected"}, "amended_design_fix": {"type": "string", "title": "Amended Design Fix"}}, "type": "object", "required": ["original_element", "failure_mode_detected", "amended_design_fix"], "title": "BlueprintRevision"}, "ChatInterrogationRequest": {"properties": {"project_id": {"type": "string", "title": "Project Id", "description": "The unique identifier of the target project domain."}, "persona_system_instruction": {"type": "string", "title": "Persona System Instruction", "description": "The explicit anthropomorphic system context generated for this specific citizen."}, "user_message": {"type": "string", "title": "User Message", "description": "The active engineering query or adjustment pitched by the city official."}, "chat_history": {"items": {"additionalProperties": {"type": "string"}, "type": "object"}, "type": "array", "title": "Chat History", "description": "The multi-turn conversational log array structured as [{'role': 'user'|'model', 'content': '...'}]", "default": []}}, "type": "object", "required": ["project_id", "persona_system_instruction", "user_message"], "title": "ChatInterrogationRequest"}, "ChatInterrogationResponse": {"properties": {"reply": {"type": "string", "title": "Reply", "description": "The first-person qualitative pushback or agreement from the citizen persona."}}, "type": "object", "required": ["reply"], "title": "ChatInterrogationResponse"}, "CitizenPersona": {"properties": {"name": {"type": "string", "title": "Name"}, "demographic_role": {"type": "string", "title": "Demographic Role"}, "sentiment_score": {"type": "integer", "title": "Sentiment Score", "description": "0 to 100 score of how much they favor the project"}, "system_instruction": {"type": "string", "title": "System Instruction", "description": "First-person prompt detailing their life, concerns, and stance for subsequent chat engagement"}}, "type": "object", "required": ["name", "demographic_role", "sentiment_score", "system_instruction"], "title": "CitizenPersona"}, "CitizenPersonaProfile": {"properties": {"name": {"type": "string", "title": "Name"}, "type": {"type": "string", "title": "Type"}, "system_instruction": {"type": "string", "title": "System Instruction"}}, "type": "object", "required": ["name", "type", "system_instruction"], "title": "CitizenPersonaProfile"}, "ConceptRenderRequest": {"properties": {"project_id": {"type": "string", "title": "Project Id"}, "version_tag": {"type": "string", "title": "Version Tag"}, "design_element_description": {"type": "string", "title": "Design Element Description"}, "environmental_context": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Environmental Context", "default": "Daytime, clean modern architecture, South Asian metropolitan context"}}, "type": "object", "required": ["project_id", "version_tag", "design_element_description"], "title": "ConceptRenderRequest"}, "ConceptRenderResponse": {"properties": {"project_id": {"type": "string", "title": "Project Id"}, "version_tag": {"type": "string", "title": "Version Tag"}, "element_rendered": {"type": "string", "title": "Element Rendered"}, "generated_image_url": {"type": "string", "title": "Generated Image Url"}, "revised_prompt_used": {"type": "string", "title": "Revised Prompt Used"}}, "type": "object", "required": ["project_id", "version_tag", "element_rendered", "generated_image_url", "revised_prompt_used"], "title": "ConceptRenderResponse"}, "FeasibilityScores": {"properties": {"social_impact_score": {"type": "number", "title": "Social Impact Score"}, "economic_viability_score": {"type": "number", "title": "Economic Viability Score"}, "political_feasibility_score": {"type": "number", "title": "Political Feasibility Score"}}, "type": "object", "required": ["social_impact_score", "economic_viability_score", "political_feasibility_score"], "title": "FeasibilityScores"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "LiveTickerMessage": {"properties": {"timestamp": {"type": "string", "title": "Timestamp"}, "agent_profile": {"type": "string", "title": "Agent Profile"}, "log_level": {"type": "string", "title": "Log Level"}, "message": {"type": "string", "title": "Message"}}, "type": "object", "required": ["timestamp", "agent_profile", "log_level", "message"], "title": "LiveTickerMessage"}, "ProjectAmendmentRequest": {"properties": {"project_id": {"type": "string", "title": "Project Id"}, "version_tag": {"type": "string", "title": "Version Tag"}, "amended_proposal_text": {"type": "string", "title": "Amended Proposal Text"}}, "type": "object", "required": ["project_id", "version_tag", "amended_proposal_text"], "title": "ProjectAmendmentRequest"}, "SimulationStateResponse": {"properties": {"project_id": {"type": "string", "title": "Project Id"}, "personas": {"items": {"$ref": "#/components/schemas/CitizenPersona"}, "type": "array", "title": "Personas"}, "social_feasibility_score": {"type": "integer", "title": "Social Feasibility Score"}, "economic_viability_score": {"type": "integer", "title": "Economic Viability Score"}, "political_acceptance_score": {"type": "integer", "title": "Political Acceptance Score"}, "arbitrator_verdict": {"type": "string", "title": "Arbitrator Verdict"}, "status": {"type": "string", "title": "Status", "default": "completed"}, "scores": {"anyOf": [{"$ref": "#/components/schemas/FeasibilityScores"}, {"type": "null"}]}, "summary_verdict": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Summary Verdict"}, "discovered_personas": {"anyOf": [{"items": {"$ref": "#/components/schemas/CitizenPersonaProfile"}, "type": "array"}, {"type": "null"}], "title": "Discovered Personas"}, "live_debate_ticks": {"anyOf": [{"items": {"$ref": "#/components/schemas/LiveTickerMessage"}, "type": "array"}, {"type": "null"}], "title": "Live Debate Ticks"}, "blueprint_revisions": {"anyOf": [{"items": {"$ref": "#/components/schemas/BlueprintRevision"}, "type": "array"}, {"type": "null"}], "title": "Blueprint Revisions"}, "spatial_telemetry": {"anyOf": [{"additionalProperties": true, "type": "object"}, {"type": "null"}], "title": "Spatial Telemetry"}}, "type": "object", "required": ["project_id", "personas", "social_feasibility_score", "economic_viability_score", "political_acceptance_score", "arbitrator_verdict"], "title": "SimulationStateResponse"}, "UnstructuredIngestRequest": {"properties": {"user_prompt": {"type": "string", "title": "User Prompt", "description": "The raw, unstructured textual layout description from the user."}}, "type": "object", "required": ["user_prompt"], "title": "UnstructuredIngestRequest"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}